relocate components to components folder
