+++ close-buy/src/App.jsx
@@ -2,11 +2,11 @@ import React, { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import axios from 'axios';
 import './App.css';
-import Splash from './Splash';
-import Search from './Search';
-import Navbar from './Navbar';
+import Splash from './components/Splash';
+import Search from './components/Search';
+import Navbar from './components/Navbar';
 import About from './components/About';
-import ProductCard from './ProductCard';
+import ProductCard from './components/ProductCard';
 import ProductView from './components/ProductView';
 import { filterData } from './helpers/selectors';
 import "slick-carousel/slick/slick.css";
