+++ close-buy/src/components/Navbar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import './Navbar.css';
-import logo from './img/closebuy-logo-white.png';
+import logo from '../img/closebuy-logo-white.png';
 
 function Navbar() {
