+++ close-buy/src/components/Search.css
