+++ close-buy/src/components/Search.jsx
