+++ close-buy/src/components/Splash.css
