+++ close-buy/src/components/Splash.jsx
